feat: 综合搜索（其实只有用户/剧集/视频）功能

+++ app/src/main/java/cn/spacexc/wearbili/dataclass/dynamic/Bangumi.kt
@@ -1,6 +1,6 @@
 package cn.spacexc.wearbili.dataclass.dynamic
 
 data class Bangumi(
-    val season_id: Int,
+    val season_id: Long,
     val type: Int
 )
