+++ app/src/main/java/cn/spacexc/wearbili/dataclass/dynamic/Desc.kt
@@ -6,7 +6,7 @@ data class Desc(
     val comment: Int?,
     val dynamic_id: Long,
     val dynamic_id_str: String,
-    val inner_id: Int?,
+    val inner_id: Long?,
     val is_liked: Int,
     val like: Long,
     val orig_dy_id: Long?,
