+++ app/src/main/java/cn/spacexc/wearbili/dataclass/dynamic/EmojiDetail.kt
@@ -3,10 +3,10 @@ package cn.spacexc.wearbili.dataclass.dynamic
 data class EmojiDetail(
     val attr: Int,
     val emoji_name: String,
-    val id: Int,
+    val id: Long,
     val meta: Meta,
     val mtime: Int,
-    val package_id: Int,
+    val package_id: Long,
     val state: Int,
     val text: String,
     val type: Int,
