+++ app/src/main/java/cn/spacexc/wearbili/dataclass/dynamic/SeasonInfo.kt
@@ -3,6 +3,6 @@ package cn.spacexc.wearbili.dataclass.dynamic
 data class SeasonInfo(
     val color: String,
     val font: String,
-    val season_id: Int,
+    val season_id: Long,
     val text: String
 )
