+++ app/src/main/java/cn/spacexc/wearbili/dataclass/dynamic/dynamicepisode/ApiSeasonInfo.kt
@@ -4,7 +4,7 @@ data class ApiSeasonInfo(
     val bgm_type: Int,
     val cover: String,
     val is_finish: Int,
-    val season_id: Int,
+    val season_id: Long,
     val title: String,
     val total_count: Int,
     val ts: Int,
