+++ app/src/main/java/cn/spacexc/wearbili/dataclass/dynamic/dynamicforwardshare/Desc.kt
@@ -5,7 +5,7 @@ data class Desc(
     val comment: Int,
     val dynamic_id: Long,
     val dynamic_id_str: String,
-    val inner_id: Int,
+    val inner_id: Long,
     val is_liked: Int,
     val like: Int,
     val orig_dy_id: Long,
