+++ app/src/main/java/cn/spacexc/wearbili/dataclass/dynamic/dynamicforwardshare/TopicDetail.kt
@@ -2,7 +2,7 @@ package cn.spacexc.wearbili.dataclass.dynamic.dynamicforwardshare
 
 data class TopicDetail(
     val is_activity: Int,
-    val topic_id: Int,
+    val topic_id: Long,
     val topic_link: String,
     val topic_name: String
 )
