+++ app/src/main/java/cn/spacexc/wearbili/dataclass/dynamic/dynamictext/card/Item.kt
@@ -4,7 +4,7 @@ data class Item(
     val content: String,
     val ctrl: String,
     val orig_dy_id: Long,
-    val pre_dy_id: Int,
+    val pre_dy_id: Long,
     val reply: Int,
     val rp_id: Long,
     val timestamp: Int,
