+++ app/src/main/java/cn/spacexc/wearbili/dataclass/dynamic/dynamicvideo/card/VideoCard.kt
@@ -3,7 +3,7 @@ package cn.spacexc.wearbili.dataclass.dynamic.dynamicvideo.card
 data class VideoCard(
     val aid: Long,
     val attribute: Int,
-    val cid: Int,
+    val cid: Long,
     val copyright: Int,
     val ctime: Int,
     val desc: String,
@@ -12,7 +12,7 @@ data class VideoCard(
     val `dynamic`: String?,
     val first_frame: String,
     val jump_url: String,
-    val mission_id: Int,
+    val mission_id: Long,
     val owner: Owner,
     val pic: String,
     val player_info: Any?,
@@ -22,7 +22,7 @@ data class VideoCard(
     val short_link_v2: String,
     val stat: Stat,
     val state: Int,
-    val tid: Int,
+    val tid: Long,
     val title: String,
     val tname: String,
     val videos: Int
