+++ app/src/main/java/cn/spacexc/wearbili/dataclass/emoji/MetaX.kt
@@ -1,6 +1,6 @@
 package cn.spacexc.wearbili.dataclass.emoji
 
 data class MetaX(
-    val item_id: Int,
+    val item_id: Long,
     val size: Int
 )
