+++ app/src/main/java/cn/spacexc/wearbili/dataclass/emoji/Setting.kt
@@ -2,6 +2,6 @@ package cn.spacexc.wearbili.dataclass.emoji
 
 data class Setting(
     val attr: Int,
-    val focus_pkg_id: Int,
+    val focus_pkg_id: Long,
     val recent_limit: Int
 )
