+++ app/src/main/java/cn/spacexc/wearbili/dataclass/follow/Group.kt
@@ -3,6 +3,6 @@ package cn.spacexc.wearbili.dataclass.follow
 data class Group(
     val count: Int,
     val name: String,
-    val tagid: Int,
+    val tagid: Long,
     val tip: String
 )
