+++ app/src/main/java/cn/spacexc/wearbili/dataclass/history/HistoryX.kt
@@ -3,10 +3,10 @@ package cn.spacexc.wearbili.dataclass.history
 data class HistoryX(
     val business: String,
     val bvid: String?,
-    val cid: Int,
+    val cid: Long,
     val dt: Int,
-    val epid: Int,
-    val oid: Int,
+    val epid: Long,
+    val oid: Long,
     val page: Int,
     val part: String
 )
