+++ app/src/main/java/cn/spacexc/wearbili/dataclass/hot/rankinglist/Premiere.kt
@@ -1,7 +1,7 @@
 package cn.spacexc.wearbili.dataclass.hot.rankinglist
 
 data class Premiere(
-    val room_id: Int,
+    val room_id: Long,
     val start_time: Int,
     val state: Int
 )
