+++ app/src/main/java/cn/spacexc/wearbili/dataclass/star/Ugc.kt
@@ -1,5 +1,5 @@
 package cn.spacexc.wearbili.dataclass.star
 
 data class Ugc(
-    val first_cid: Int
+    val first_cid: Long
 )
