+++ app/src/main/java/cn/spacexc/wearbili/dataclass/video/DescV2.kt
@@ -1,7 +1,7 @@
 package cn.spacexc.wearbili.dataclass.video
 
 data class DescV2(
-    val biz_id: Int,
+    val biz_id: Long,
     val raw_text: String,
     val type: Int
 )
