+++ app/src/main/java/cn/spacexc/wearbili/dataclass/video/Episode.kt
@@ -1,14 +1,14 @@
 package cn.spacexc.wearbili.dataclass.video
 
 data class Episode(
-    val aid: Int,
+    val aid: Long,
     val arc: Arc,
     val attribute: Int,
     val bvid: String,
-    val cid: Int,
-    val id: Int,
+    val cid: Long,
+    val id: Long,
     val page: PageX,
-    val season_id: Int,
-    val section_id: Int,
+    val season_id: Long,
+    val section_id: Long,
     val title: String
 )
