+++ app/src/main/java/cn/spacexc/wearbili/dataclass/video/Honor.kt
@@ -1,7 +1,7 @@
 package cn.spacexc.wearbili.dataclass.video
 
 data class Honor(
-    val aid: Int,
+    val aid: Long,
     val desc: String,
     val type: Int,
     val weekly_recommend_num: Int
