+++ app/src/main/java/cn/spacexc/wearbili/dataclass/video/Section.kt
@@ -2,8 +2,8 @@ package cn.spacexc.wearbili.dataclass.video
 
 data class Section(
     val episodes: List<Episode>,
-    val id: Int,
-    val season_id: Int,
+    val id: Long,
+    val season_id: Long,
     val title: String,
     val type: Int
 )
