+++ app/src/main/java/cn/spacexc/wearbili/dataclass/video/rcmd/Args.kt
@@ -1,10 +1,10 @@
 package cn.spacexc.wearbili.dataclass.video.rcmd
 
 data class Args(
-    val aid: Int,
-    val rid: Int,
+    val aid: Long,
+    val rid: Long,
     val rname: String,
-    val tid: Int?,
+    val tid: Long?,
     val tname: String?,
     val up_id: Long,
     val up_name: String?
