+++ app/src/main/java/cn/spacexc/wearbili/dataclass/video/rcmd/FeedbackPanelDetail.kt
@@ -4,7 +4,7 @@ data class FeedbackPanelDetail(
     val icon_url: String,
     val jump_type: Int,
     val jump_url: String,
-    val module_id: Int,
+    val module_id: Long,
     val secondary_panel: List<SecondaryPanel>?,
     val sub_text: String,
     val text: String
