+++ app/src/main/java/cn/spacexc/wearbili/dataclass/video/rcmd/SecondaryPanel.kt
@@ -1,6 +1,6 @@
 package cn.spacexc.wearbili.dataclass.video.rcmd
 
 data class SecondaryPanel(
-    val reason_id: Int,
+    val reason_id: Long,
     val text: String
 )
