+++ app/src/main/java/cn/spacexc/wearbili/dataclass/videoDetail/Reason.kt
@@ -1,9 +1,9 @@
 package cn.spacexc.wearbili.dataclass.videoDetail
 
 data class Reason(
-    val id: Int,
+    val id: Long,
     val mid: Long,
     val name: String,
-    val rid: Int,
-    val tag_id: Int
+    val rid: Long,
+    val tag_id: Long
 )
